+++ tests/functional/server/chalk.py
@@ -3,6 +3,7 @@
 # This file is part of Chalk
 # (see https://crashoverride.com/docs/chalk)
 import asyncio
+import hashlib
 import json
 import os
 import pathlib
@@ -66,15 +67,35 @@ async def ping(stats: list[schemas.Stat], db: Session = Depends(get_db)):
         return {"ping": "pong"}
 
 
+async def _check_chalk_core_headers(
+    request: Request, verify_body: bool = False
+) -> None:
+    for header in [
+        "x-chalk-version",
+        "x-chalk-operation",
+        "x-chalk-action-id",
+        "x-content-length",
+        "x-chalk-digest-sha256",
+    ]:
+        if not request.headers.get(header):
+            raise HTTPException(status_code=400, detail=f"missing {header} header")
+    if verify_body:
+        body = await request.body()
+        if int(request.headers["x-content-length"]) != len(body):
+            raise HTTPException(status_code=400, detail="x-content-length mismatch")
+        expected = hashlib.sha256(body).hexdigest()
+        if request.headers["x-chalk-digest-sha256"] != expected:
+            raise HTTPException(
+                status_code=400, detail="x-chalk-digest-sha256 mismatch"
+            )
+
+
 @app.put("/report/presign", status_code=200)
 async def presign_report_url(
     request: Request,
     response: Response,
 ):
-    if not request.headers.get("x-chalk-version"):
-        raise HTTPException(status_code=400, detail="missing X-Chalk-Version header")
-    if not request.headers.get("x-chalk-action-id"):
-        raise HTTPException(status_code=400, detail="missing X-Chalk-Action-Id header")
+    await _check_chalk_core_headers(request)
     redirect = RedirectResponse(
         request.url_for("accept_presign_report"),
         status_code=status.HTTP_307_TEMPORARY_REDIRECT,
@@ -115,14 +136,7 @@ async def accept_report(
     db: Session = Depends(get_db),
 ):
     if request.method == "POST":
-        if not request.headers.get("x-chalk-version"):
-            raise HTTPException(
-                status_code=400, detail="missing X-Chalk-Version header"
-            )
-        if not request.headers.get("x-chalk-action-id"):
-            raise HTTPException(
-                status_code=400, detail="missing X-Chalk-Action-Id header"
-            )
+        await _check_chalk_core_headers(request, verify_body=True)
     try:
         for report in reports:
             operation = report.get("_OPERATION")